Substitution: fixes and tests.
Mathsat5 native API had linking issues, a potential TODO.
This change makes types more robust.

+++ src/org/sosy_lab/solver/mathsat5/Mathsat5NativeApi.java
@@ -455,7 +455,7 @@ class Mathsat5NativeApi {
 
   public static native long msat_make_copy_from(long e, long t, long src);
 
-  public static native long msat_apply_substitution(long e, long t, long s, long[] from, long[] to);
+  public static native long msat_apply_substitution(long e, long t, int s, long[] from, long[] to);
 
   /*
    * Term access and navigation
